Nit: silence -Wshorten-64-to-32 on the DECIMAL precision calculation (#1475)
field->length is unsigned long (64-bit); precision is int (32-bit).
DECIMAL's max precision is 65 digits, so field->length for any
legitimate DECIMAL column never exceeds ~70 -- the narrowing is safe,
just unproven to the compiler. Add an explicit cast to document that
and quiet the warning, which only appeared on Apple Clang builds.

+++ ext/mysql2/result.c
@@ -313,7 +313,8 @@ static VALUE rb_mysql_result_fetch_field_type(VALUE self, unsigned int idx) {
           Handle precision similar to this line from mysql's code:
           https://github.com/mysql/mysql-server/blob/ea7d2e2d16ac03afdd9cb72a972a95981107bf51/sql/field.cc#L2246
         */
-        precision = field->length - (field->decimals > 0 ? 2 : 1);
+        // DECIMAL's max precision is 65 digits, so this narrowing is safe for any field the server actually sent.
+        precision = (int)(field->length - (field->decimals > 0 ? 2 : 1));
         rb_field_type = rb_sprintf("decimal(%d,%d)", precision, field->decimals);
         break;
       case MYSQL_TYPE_STRING:       // char[]
